Making general changes
Made changes to the CSS for the Buttons, Navbar, and Dropdown Menus so everything is working as it should. Also Changed the screen size when the navbar gets display of none, did this because I wanted more spacing between the nav items. Added js for the other websites link and button so the button works to open and close the other websites dropdown menu and so when the page loads the dropdown is hidden. Also needed to move the navigation bar margin to the left so it wasn't overlapping the button, which is why the button yesterday wasn't working, so I am glad I solved that issue lol.