General changes to css html and js files
Adding an id to the dropdown-btn so I can better target it with the new JavaScript function, speaking of which I deleted the old function that was supposed to close the other websites drop down nav but didn't. I replaced it with a new, it seems to work so lets see when I push it.  Then for the CSS changes I moved the onpage mobile nav dropdown menu to center it with the button. Then I changed up the padding with the list items in side of the mobile dropdown menu and other sites menu